server/requirements.txt
Django>=1.11
django-filter==1.0.2
djangorestframework==3.6.2
dj-database-url==0.3.0
python-social-auth==0.3.6
social-auth-app-django==1.1.0
six>=1.8.0
uwsgi==2.0.15
praw==4.5.1
